refactor: Rewrite webdav to improve code quality
