+++ core/src/services/webdav/backend.rs
@@ -16,22 +16,18 @@
 // under the License.
 
 use std::collections::HashMap;
-use std::collections::VecDeque;
 use std::fmt::Debug;
 use std::fmt::Formatter;
+use std::str::FromStr;
+use std::sync::Arc;
 
 use async_trait::async_trait;
-use bytes::Buf;
-use http::header;
-use http::HeaderMap;
-use http::Request;
-use http::Response;
 use http::StatusCode;
 use log::debug;
 use serde::Deserialize;
 
+use super::core::*;
 use super::error::parse_error;
-use super::lister::Multistatus;
 use super::lister::WebdavLister;
 use super::writer::WebdavWriter;
 use crate::raw::*;
@@ -177,6 +173,16 @@ impl Builder for WebdavBuilder {
                     .with_context("service", Scheme::Webdav));
             }
         };
+        // Some services might return the path with suffix `/remote.php/webdav/`, we need to trim them.
+        let server_path = http::Uri::from_str(endpoint)
+            .map_err(|err| {
+                Error::new(ErrorKind::ConfigInvalid, "endpoint is invalid")
+                    .with_context("service", Scheme::Webdav)
+                    .set_source(err)
+            })?
+            .path()
+            .trim_end_matches('/')
+            .to_string();
 
         let root = normalize_root(&self.config.root.clone().unwrap_or_default());
         debug!("backend use root {}", root);
@@ -190,45 +196,41 @@ impl Builder for WebdavBuilder {
             })?
         };
 
-        let mut auth = None;
+        let mut authorization = None;
         if let Some(username) = &self.config.username {
-            auth = Some(format_authorization_by_basic(
+            authorization = Some(format_authorization_by_basic(
                 username,
                 self.config.password.as_deref().unwrap_or_default(),
             )?);
         }
         if let Some(token) = &self.config.token {
-            auth = Some(format_authorization_by_bearer(token)?)
+            authorization = Some(format_authorization_by_bearer(token)?)
         }
 
         debug!("backend build finished: {:?}", &self);
-        Ok(WebdavBackend {
+
+        let core = Arc::new(WebdavCore {
             endpoint: endpoint.to_string(),
-            authorization: auth,
+            server_path,
+            authorization,
             disable_copy: self.config.disable_copy,
             root,
             client,
-        })
+        });
+        Ok(WebdavBackend { core })
     }
 }
 
 /// Backend is used to serve `Accessor` support for http.
 #[derive(Clone)]
 pub struct WebdavBackend {
-    endpoint: String,
-    root: String,
-    client: HttpClient,
-    disable_copy: bool,
-
-    authorization: Option<String>,
+    core: Arc<WebdavCore>,
 }
 
 impl Debug for WebdavBackend {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Backend")
-            .field("endpoint", &self.endpoint)
-            .field("root", &self.root)
-            .field("client", &self.client)
+        f.debug_struct("WebdavBackend")
+            .field("core", &self.core)
             .finish()
     }
 }
@@ -237,7 +239,7 @@ impl Debug for WebdavBackend {
 impl Accessor for WebdavBackend {
     type Reader = IncomingAsyncBody;
     type Writer = oio::OneShotWriter<WebdavWriter>;
-    type Lister = Option<oio::PageLister<WebdavLister>>;
+    type Lister = oio::PageLister<WebdavLister>;
     type BlockingReader = ();
     type BlockingWriter = ();
     type BlockingLister = ();
@@ -245,7 +247,7 @@ impl Accessor for WebdavBackend {
     fn info(&self) -> AccessorInfo {
         let mut ma = AccessorInfo::default();
         ma.set_scheme(Scheme::Webdav)
-            .set_root(&self.root)
+            .set_root(&self.core.root)
             .set_native_capability(Capability {
                 stat: true,
 
@@ -259,7 +261,7 @@ impl Accessor for WebdavBackend {
                 create_dir: true,
                 delete: true,
 
-                copy: !self.disable_copy,
+                copy: !self.core.disable_copy,
 
                 rename: true,
 
@@ -272,60 +274,17 @@ impl Accessor for WebdavBackend {
     }
 
     async fn create_dir(&self, path: &str, _: OpCreateDir) -> Result<RpCreateDir> {
-        self.ensure_parent_path(path).await?;
-        self.create_dir_internal(path).await?;
-
+        self.core.webdav_mkcol(path).await?;
         Ok(RpCreateDir::default())
     }
 
     async fn stat(&self, path: &str, _: OpStat) -> Result<RpStat> {
-        let mut header_map = HeaderMap::new();
-        // not include children
-        header_map.insert("Depth", "0".parse().unwrap());
-        header_map.insert(header::ACCEPT, "application/xml".parse().unwrap());
-
-        let resp = self.webdav_propfind(path, Some(header_map)).await?;
-
-        let status = resp.status();
-
-        if !status.is_success() {
-            return Err(parse_error(resp).await?);
-        }
-
-        let bs = resp.into_body().bytes().await?;
-        let s = String::from_utf8_lossy(&bs);
-
-        // Make sure the string is escaped.
-        // Related to <https://github.com/tafia/quick-xml/issues/719>
-        //
-        // This is a temporary solution, we should find a better way to handle this.
-        let s = s.replace("&()_+-=;", "%26%28%29_%2B-%3D%3B");
-        let result: Multistatus = quick_xml::de::from_str(&s).map_err(new_xml_deserialize_error)?;
-
-        let response = match result.response {
-            Some(v) => v,
-            None => {
-                return Err(Error::new(
-                    ErrorKind::NotFound,
-                    "Failed getting item stat: response field was not found",
-                ));
-            }
-        };
-
-        let item = response
-            .first()
-            .ok_or_else(|| {
-                Error::new(
-                    ErrorKind::Unexpected,
-                    "Failed getting item stat: bad response",
-                )
-            })?
-            .parse_into_metadata()?;
-        Ok(RpStat::new(item))
+        let metadata = self.core.webdav_stat(path).await?;
+        Ok(RpStat::new(metadata))
     }
 
     async fn read(&self, path: &str, args: OpRead) -> Result<(RpRead, Self::Reader)> {
-        let resp = self.webdav_get(path, args).await?;
+        let resp = self.core.webdav_get(path, args).await?;
         let status = resp.status();
         match status {
             StatusCode::OK | StatusCode::PARTIAL_CONTENT => {
@@ -345,21 +304,19 @@ impl Accessor for WebdavBackend {
     }
 
     async fn write(&self, path: &str, args: OpWrite) -> Result<(RpWrite, Self::Writer)> {
-        self.ensure_parent_path(path).await?;
-
-        let p = build_abs_path(&self.root, path);
+        // Ensure parent path exists
+        self.core.webdav_mkcol(get_parent(path)).await?;
 
         Ok((
             RpWrite::default(),
-            oio::OneShotWriter::new(WebdavWriter::new(self.clone(), args, p)),
+            oio::OneShotWriter::new(WebdavWriter::new(self.core.clone(), args, path.to_string())),
         ))
     }
 
     async fn delete(&self, path: &str, _: OpDelete) -> Result<RpDelete> {
-        let resp = self.webdav_delete(path).await?;
+        let resp = self.core.webdav_delete(path).await?;
 
         let status = resp.status();
-
         match status {
             StatusCode::NO_CONTENT | StatusCode::NOT_FOUND => Ok(RpDelete::default()),
             _ => Err(parse_error(resp).await?),
@@ -367,57 +324,14 @@ impl Accessor for WebdavBackend {
     }
 
     async fn list(&self, path: &str, args: OpList) -> Result<(RpList, Self::Lister)> {
-        if args.recursive() {
-            return Err(Error::new(
-                ErrorKind::Unsupported,
-                "webdav doesn't support list with recursive",
-            ));
-        }
-
-        let mut header_map = HeaderMap::new();
-        header_map.insert("Depth", "1".parse().unwrap());
-        header_map.insert(header::CONTENT_TYPE, "application/xml".parse().unwrap());
-        let resp = self.webdav_propfind(path, Some(header_map)).await?;
-        let status = resp.status();
-
-        match status {
-            StatusCode::OK | StatusCode::MULTI_STATUS => {
-                let bs = resp.into_body().bytes().await?;
-                let result: Multistatus =
-                    quick_xml::de::from_reader(bs.reader()).map_err(new_xml_deserialize_error)?;
-
-                let result = match result.response {
-                    Some(v) => v,
-                    None => {
-                        return Ok((RpList::default(), None));
-                    }
-                };
-
-                let l = WebdavLister::new(&self.endpoint, &self.root, path, result);
-
-                Ok((RpList::default(), Some(oio::PageLister::new(l))))
-            }
-            StatusCode::NOT_FOUND if path.ends_with('/') => Ok((RpList::default(), None)),
-            _ => Err(parse_error(resp).await?),
-        }
+        Ok((
+            RpList::default(),
+            oio::PageLister::new(WebdavLister::new(self.core.clone(), path, args)),
+        ))
     }
 
-    /// # Notes
-    ///
-    /// There is a strange dead lock issues when copying a non-exist file, so we will check
-    /// if the source exists first.
-    ///
-    /// For example: <https://github.com/apache/opendal/pull/2809>
     async fn copy(&self, from: &str, to: &str, _args: OpCopy) -> Result<RpCopy> {
-        if let Err(err) = self.stat(from, OpStat::default()).await {
-            if err.kind() == ErrorKind::NotFound {
-                return Err(err);
-            }
-        }
-
-        self.ensure_parent_path(to).await?;
-
-        let resp = self.webdav_copy(from, to).await?;
+        let resp = self.core.webdav_copy(from, to).await?;
 
         let status = resp.status();
 
@@ -428,293 +342,14 @@ impl Accessor for WebdavBackend {
     }
 
     async fn rename(&self, from: &str, to: &str, _args: OpRename) -> Result<RpRename> {
-        self.ensure_parent_path(to).await?;
-
-        let resp = self.webdav_move(from, to).await?;
+        let resp = self.core.webdav_move(from, to).await?;
 
         let status = resp.status();
-
         match status {
             StatusCode::CREATED | StatusCode::NO_CONTENT | StatusCode::OK => {
                 Ok(RpRename::default())
             }
             _ => Err(parse_error(resp).await?),
         }
-    }
-}
-
-impl WebdavBackend {
-    async fn webdav_get(&self, path: &str, args: OpRead) -> Result<Response<IncomingAsyncBody>> {
-        let p = build_rooted_abs_path(&self.root, path);
-        let url: String = format!("{}{}", self.endpoint, percent_encode_path(&p));
-
-        let mut req = Request::get(&url);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth.clone())
-        }
-
-        let range = args.range();
-        if !range.is_full() {
-            req = req.header(header::RANGE, range.to_header());
-        }
-
-        let req = req
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    pub async fn webdav_put(
-        &self,
-        abs_path: &str,
-        size: Option<u64>,
-        args: &OpWrite,
-        body: AsyncBody,
-    ) -> Result<Response<IncomingAsyncBody>> {
-        let url = format!("{}/{}", self.endpoint, percent_encode_path(abs_path));
-
-        let mut req = Request::put(&url);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth.clone())
-        }
-
-        if let Some(size) = size {
-            req = req.header(header::CONTENT_LENGTH, size)
-        }
-
-        if let Some(mime) = args.content_type() {
-            req = req.header(header::CONTENT_TYPE, mime)
-        }
-
-        if let Some(cd) = args.content_disposition() {
-            req = req.header(header::CONTENT_DISPOSITION, cd)
-        }
-
-        // Set body
-        let req = req.body(body).map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn webdav_mkcol_absolute_path(&self, path: &str) -> Result<Response<IncomingAsyncBody>> {
-        debug_assert!(path.starts_with('/'), "path must be absolute path");
-        let url = format!("{}{}", self.endpoint, percent_encode_path(path));
-
-        let mut req = Request::builder().method("MKCOL").uri(&url);
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth);
-        }
-
-        let req = req
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn webdav_propfind(
-        &self,
-        path: &str,
-        headers: Option<HeaderMap>,
-    ) -> Result<Response<IncomingAsyncBody>> {
-        let p = build_rooted_abs_path(&self.root, path);
-
-        self.webdav_propfind_absolute_path(&p, headers).await
-    }
-
-    async fn webdav_propfind_absolute_path(
-        &self,
-        path: &str,
-        headers: Option<HeaderMap>,
-    ) -> Result<Response<IncomingAsyncBody>> {
-        debug_assert!(path.starts_with('/'), "path must be absolute path");
-
-        let url = format!("{}{}", self.endpoint, percent_encode_path(path));
-        let mut req = Request::builder().method("PROPFIND").uri(&url);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth);
-        }
-
-        if let Some(headers) = headers {
-            for (name, value) in headers {
-                // all key should be not None, otherwise panic
-                req = req.header(name.unwrap(), value);
-            }
-        }
-
-        // rfc4918 9.1: retrieve all properties define in specification
-        let body;
-        {
-            req = req.header(header::CONTENT_TYPE, "application/xml");
-            // XML body must start without a new line. Otherwise, the server will panic: `xmlParseChunk() failed`
-            let all_prop_xml_body = r#"<?xml version="1.0" encoding="utf-8" ?>
-            <D:propfind xmlns:D="DAV:">
-                <D:allprop/>
-            </D:propfind>
-        "#;
-            body = AsyncBody::Bytes(bytes::Bytes::from(all_prop_xml_body));
-        }
-
-        let req = req.body(body).map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn webdav_delete(&self, path: &str) -> Result<Response<IncomingAsyncBody>> {
-        let p = build_abs_path(&self.root, path);
-
-        let url = format!("{}/{}", self.endpoint, percent_encode_path(&p));
-
-        let mut req = Request::delete(&url);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth.clone())
-        }
-
-        let req = req
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn webdav_copy(&self, from: &str, to: &str) -> Result<Response<IncomingAsyncBody>> {
-        let source = build_abs_path(&self.root, from);
-        let target = build_abs_path(&self.root, to);
-        // Make sure target's dir is exist.
-        self.ensure_parent_path(&target).await?;
-
-        let source = format!("{}/{}", self.endpoint, percent_encode_path(&source));
-        let target = format!("{}/{}", self.endpoint, percent_encode_path(&target));
-
-        let mut req = Request::builder().method("COPY").uri(&source);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth);
-        }
-
-        req = req.header("Destination", target);
-
-        // We always specific "T" for keeping to overwrite the destination.
-        req = req.header("Overwrite", "T");
-
-        let req = req
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn webdav_move(&self, from: &str, to: &str) -> Result<Response<IncomingAsyncBody>> {
-        // Check if the source exists first.
-        self.stat(from, OpStat::new()).await?;
-
-        let source = build_abs_path(&self.root, from);
-        let target = build_abs_path(&self.root, to);
-        // Make sure target's dir is exist.
-        self.ensure_parent_path(&target).await?;
-
-        let source = format!("{}/{}", self.endpoint, percent_encode_path(&source));
-        let target = format!("{}/{}", self.endpoint, percent_encode_path(&target));
-
-        let mut req = Request::builder().method("MOVE").uri(&source);
-
-        if let Some(auth) = &self.authorization {
-            req = req.header(header::AUTHORIZATION, auth);
-        }
-
-        req = req.header("Destination", target);
-
-        // We always specific "T" for keeping to overwrite the destination.
-        req = req.header("Overwrite", "T");
-
-        let req = req
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-
-        self.client.send(req).await
-    }
-
-    async fn create_dir_internal(&self, path: &str) -> Result<()> {
-        let p = build_rooted_abs_path(&self.root, path);
-        self.create_dir_internal_absolute_path(&p).await
-    }
-
-    async fn create_dir_internal_absolute_path(&self, path: &str) -> Result<()> {
-        debug_assert!(path.starts_with('/'), "path must be absolute path");
-
-        let resp = self.webdav_mkcol_absolute_path(path).await?;
-
-        let status = resp.status();
-
-        match status {
-            StatusCode::CREATED
-            // Allow multiple status
-            | StatusCode::MULTI_STATUS
-            // The MKCOL method can only be performed on a deleted or non-existent resource.
-            // This error means the directory already exists which is allowed by create_dir.
-            | StatusCode::METHOD_NOT_ALLOWED => {
-                resp.into_body().consume().await?;
-                Ok(())
-            }
-            _ => Err(parse_error(resp).await?),
-        }
-    }
-
-    async fn ensure_parent_path(&self, path: &str) -> Result<()> {
-        let path = build_rooted_abs_path(&self.root, path);
-        let mut path = path.as_str();
-
-        let mut dirs = VecDeque::default();
-
-        loop {
-            // check path first.
-            let parent = get_parent(path);
-
-            let mut header_map = HeaderMap::new();
-            // not include children
-            header_map.insert("Depth", "0".parse().unwrap());
-            header_map.insert(header::ACCEPT, "application/xml".parse().unwrap());
-
-            let resp = self
-                .webdav_propfind_absolute_path(parent, Some(header_map))
-                .await?;
-            match resp.status() {
-                StatusCode::OK => {
-                    break;
-                }
-                StatusCode::MULTI_STATUS => {
-                    let bs = resp.into_body().bytes().await?;
-                    let s = String::from_utf8_lossy(&bs);
-                    let result: Multistatus =
-                        quick_xml::de::from_str(&s).map_err(new_xml_deserialize_error)?;
-
-                    if result.response.is_some() {
-                        break;
-                    }
-
-                    dirs.push_front(parent);
-                    path = parent
-                }
-                StatusCode::NOT_FOUND => {
-                    dirs.push_front(parent);
-                    path = parent
-                }
-                _ => return Err(parse_error(resp).await?),
-            }
-
-            if path == "/" {
-                break;
-            }
-        }
-
-        for dir in dirs {
-            self.create_dir_internal_absolute_path(dir).await?;
-        }
-        Ok(())
     }
 }
